requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-storage
python-multipart
pydantic[dotenv]
python-dotenv
pymongo==4.9.1
motor==3.6.0
fastapi-mail
passlib[bcrypt]
setuptools